+++ app-backend-atelie/src/main/java/com/eseltech/appbackendatelie/DTO/OrcamentoDTO.java
@@ -17,7 +17,7 @@ public record OrcamentoDTO(
                 example = "1",
                 accessMode = Schema.AccessMode.READ_ONLY
         )
-        Long id,
+        Integer id,
 
         @NotNull(message = "O ID da empresa é obrigatório")
         @Schema(
@@ -25,7 +25,7 @@ public record OrcamentoDTO(
                 example = "1",
                 requiredMode = Schema.RequiredMode.REQUIRED
         )
-        Long empresaId,
+        Integer empresaId,
 
         @NotNull(message = "O título é obrigatório")
         @NotBlank(message = "O título não pode estar em branco")
