+++ app-backend-atelie/src/main/java/com/eseltech/appbackendatelie/repository/MaterialProdutoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MaterialProdutoRepository extends JpaRepository<MaterialProduto, Long> {
+public interface MaterialProdutoRepository extends JpaRepository<MaterialProduto, Integer> {
 }
